requirements.txt
python-telegram-bot==22.3
gallery-dl